requirements.txt
amqp==5.1.1
asgiref==3.5.2
billiard==3.6.4.0
Brotli==1.0.9
celery==5.2.7
certifi==2022.6.15
cffi==1.15.1
charset-normalizer==2.1.0
click==8.1.3
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.2.0
colorama==0.4.5
ConfigArgParse==1.5.3
coreapi==2.3.3
coreschema==0.0.4
cryptography==37.0.4
defusedxml==0.7.1
dj-database-url==1.0.0
Django==4.1
django-cors-headers==3.13.0
django-debug-toolbar==3.5.0
django-filter==22.1
django-jazzmin==2.5.0
django-templated-mail==1.1.1
djangorestframework==3.13.1
djangorestframework-simplejwt==4.8.0
djoser==2.1.0
drf-nested-routers==0.93.4
drf-yasg==1.21.3
Flask==2.2.2
Flask-BasicAuth==0.2.0
Flask-Cors==3.0.10
flower==1.2.0
gevent==21.12.0
geventhttpclient==2.0
greenlet==1.1.2
gunicorn==20.1.0
humanize==4.3.0
idna==3.3
inflection==0.5.1
itsdangerous==2.1.2
itypes==1.2.0
Jinja2==3.1.2
kombu==5.2.4
locust==2.11.0
MarkupSafe==2.1.1
msgpack==1.0.4
mysqlclient==2.1.1
oauthlib==3.2.0
packaging==21.3
Pillow==9.2.0
prometheus-client==0.14.1
prompt-toolkit==3.0.30
psutil==5.9.1
pycparser==2.21
PyJWT==2.4.0
pyparsing==3.0.9
python3-openid==3.2.0
pytz==2022.2.1
pywin32==304
pyzmq==22.3.0
requests==2.28.1
requests-oauthlib==1.3.1
roundrobin==0.0.4
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
social-auth-app-django==4.0.0
social-auth-core==4.3.0
sqlparse==0.4.2
tornado==6.2
typing_extensions==4.3.0
tzdata==2022.2
uritemplate==4.1.1
urllib3==1.26.11
vine==5.0.0
wcwidth==0.2.5
Werkzeug==2.2.2
whitenoise==6.2.0
zope.event==4.5.0
zope.interface==5.4.0
